Resolve issue #2
This commit includes two files; one for desterminstic functions of random
permutation and tests cases with pytest.
djsolver determine whether the input is balanced or constant with the input
of a derminstic function in 2^(n-1) range.
random_solve gets random integer of 1 and 0 from size of the given n and
returns constant with a chance of 70% and balanced of 30%.
Nrandom_solve takes the numebr of k that counts the number of balanced
intergers determined by the random_solve and return balanced if it is
greater than half. Otherwise, return constant.
Tests file contains two lambda functions and tests of djsolver, solve and
random_solve.