Started on Hello Quantum Solutions
Added a folder for storing Hello Quantum Solutions. In that folder
is a README and hq-1-1.py. The later contains a CIRQ encoding of
Level 01, Puzzle 01 of Hello Quantum and it's solution. It also
provide a few different ways of running and simulating the circuit.
Other puzzle solutions needn't display the circuit's simulation
in all these ways. This is as much demo as it is anything else.