Rough start on #11 - Analysis/Visualization for HQ
Created hqhelp.py which contains two functions: pBarsStd and pBarsBell.
The former plots the probability of getting each standard basis
from a standard measurement when given a state from a cirq simulation.
The later does the same but with respect to the Bell basis.

This works OK but a better graphic might be one bar per step thereby
plotting the probability for each state at each step on a single
chart.