Updates to Analysis/Viz library for HQ Game
A bigger step towards Issues #11.  This adds a couple of key things:
a set of functions for computing state vectors and probability
vectors for all the HelloQuantum measurement combos, a function
hqGrid which plots the probability distributions of all four
measurement combinations a la the center grid in HelloQuantum, and
measurement specific functions for plot outcome probabilities for
one or more steps.

The big add here is hqGrid. This graphic lets you get a more
probabilistic approach to the game's visualization. What's needed
still are functions for tracing out individual bits and capturing
the game's visualization at the bit level.