Added a Script with a few Deutsch-Jozsa circuits
The dj_example.py script contains a function make_dj_circuit
that takes the number of bits of the boolean function f and
a unitary encoding of f (for length+1 bits) and produces
a full circuit for the Quantum Deutsch-Jozsa solver.  There are
then 3 examples: two balanced functions and one constant.