h2_example.py is now a start to finish example
The script h2_example.py is largely the example from the
openfermion-cirq tutorial with a few bug fixes (?) and
the addition of psi4.  At this point the results differ
from what you see in the tutorial some something isn't
correct.  Still, it should serve as a starting place
for exploring the parts of openfermion, openfermion-cirq, and
cirq that drive VQE simulatons of electronic structure.